+++ tests/test_fields.py
@@ -21,7 +21,7 @@ from nose.tools import eq_
 
 from tests.fixtures import BananaData, EcoProducerData, PenData, ShopData
 from tests.test_djeneralize.fruit.models import Banana
-from tests.test_djeneralize.producers.models import EcoProducer, Shop
+from tests.test_djeneralize.producers.models import EcoProducer
 from tests.test_djeneralize.writing.models import WritingImplement
 
 
@@ -34,27 +34,3 @@ class TestForeignKey(FixtureTestCase):
         eco = EcoProducer.objects.get(name=EcoProducerData.BananaProducer.name)
         eq_(eco.produce.__class__, Banana)
         eq_(eco.pen.__class__, WritingImplement)
-    
-    def test_specialized_related_manager(self):
-        """
-        The default specialization manager is used for related object lookups
-        
-        """
-        eco = EcoProducer.objects.get(name=EcoProducerData.BananaProducer.name)
-        
-        banana = Banana.objects.get(name=BananaData.Banana.name)
-        related_object = banana.fruitproducer_set.all()[0]
-
-        eq_(eco, related_object)
-    
-    def test_non_specialized_related_manager(self):
-        """
-        SpecializedForeignKey handles vanilla Django model sub-classes as well
-        
-        """
-        
-        eco = EcoProducer.objects.get(name=EcoProducerData.BananaProducer.name)
-        shop = Shop.objects.get(name=ShopData.EcoMart.name)
-        
-        related_object = eco.shops.all()[0]
-        eq_(shop, related_object)