+++ docs/source/changelog.rst
@@ -2,6 +2,13 @@
 Changelog for :mod:`djeneralize`
 ================================
 
+Version 2.0 RC1
+===============
+
+- Added support for Django 1.4.
+- Dropped specialization of objects returned by reverse relationships
+  of :class:`~djeneralize.fields.SpecializedForeignKey` fields.
+
 Version 1.1
 ===========